requirements.txt
aiogram==3.22.0
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
asyncpg==0.30.0
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0
apscheduler==3.11.2
alembic==1.18.4
psycopg2-binary
openai>=1.30.0
Pillow>=10.0.0